requirements.txt
uvicorn
fastapi
python-dotenv
pymongo
python-multipart